fix time_t formats

+++ regress/sys/kern/kqueue/write/inet/inet.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: inet.c,v 1.8 2008/04/28 20:23:06 martin Exp $	*/
+/*	$NetBSD: inet.c,v 1.9 2008/12/29 05:56:02 christos Exp $	*/
 
 /*-
  * Copyright (c) 2002 The NetBSD Foundation, Inc.
@@ -108,8 +108,8 @@ sleep(5);
 			err(1, "gettimeofday now");
 		timersub(&now, &then, &diff);
 
-		printf("iw: kevent returned %d in %ld.%06ld\n", n,
-		    diff.tv_sec, diff.tv_usec);
+		printf("iw: kevent returned %d in %lld.%06ld\n", n,
+		    (long long)diff.tv_sec, (long)diff.tv_usec);
 		if (n == -1)
 			err(1, "kevent");
 		else if (n == 0)
@@ -212,8 +212,8 @@ printf("ir: registered sock %d\n", sock);
 			err(1, "ir: gettimeofday now");
 		timersub(&now, &then, &diff);
 
-		printf("ir: kevent returned %d in %ld.%06ld\n", n,
-		    diff.tv_sec, diff.tv_usec);
+		printf("ir: kevent returned %d in %lld.%06ld\n", n,
+		    (long long)diff.tv_sec, (long)diff.tv_usec);
 		if (n == -1)
 			err(1, "ir: kevent");
 		else if (n == 0)
@@ -247,8 +247,8 @@ printf("ir: registered fd %d\n", fd);
 			err(1, "ir: gettimeofday now");
 		timersub(&now, &then, &diff);
 
-		printf("ir: kevent returned %d in %ld.%06ld\n", n,
-		    diff.tv_sec, diff.tv_usec);
+		printf("ir: kevent returned %d in %lld.%06ld\n", n,
+		    (long long)diff.tv_sec, (long)diff.tv_usec);
 		if (n == -1)
 			err(1, "ir: kevent");
 		else if (n == 0)
